refactor: 컴포넌트 common 폴더 생성
