+++ src/components/common/Header.tsx
@@ -1,13 +1,13 @@
 import styled from "styled-components";
-import { CartIcon } from "../assets";
+import { CartIcon } from "../../assets";
 import { useRecoilValue } from "recoil";
-import { cartNumberSelector } from "../recoil/selector";
-import { ROUTER_PATH } from "../router";
-import { useRouter } from "../hooks/useRouter";
-import { loginState, userState } from "../recoil/atom";
+import { cartNumberSelector } from "../../recoil/selector";
+import { ROUTER_PATH } from "../../router";
+import { useRouter } from "../../hooks/useRouter";
+import { loginState, userState } from "../../recoil/atom";
 import { useLocation } from "react-router-dom";
-import { ServerSelectBox } from "./ServerSelectBox";
-import { useLoginForm } from "../hooks/useLoginForm";
+import { ServerSelectBox } from "./../ServerSelectBox";
+import { useLoginForm } from "../../hooks/useLoginForm";
 
 export const Header = () => {
   const { goPage } = useRouter();
