+++ src/components/index.ts
@@ -1,14 +1,14 @@
-export { Counter } from "./Counter";
-export { Header } from "./Header";
-export { Page } from "./Page";
-export { Button } from "./Button";
-export { Loading } from "./Loading";
-export { ErrorBox } from "./ErrorBox";
-export { GuideBox } from "./GuideBox";
-export { Toast } from "./Toast";
-export { ErrorBoundary } from "./ErrorBoundary";
+export { Header } from "./common/Header";
+export { Page } from "./common/Page";
+export { Button } from "./common/Button";
+export { Loading } from "./common/Loading";
+export { ErrorBox } from "./common/ErrorBox";
+export { GuideBox } from "./common/GuideBox";
+export { Toast } from "./common/Toast";
+export { ErrorBoundary } from "./common/ErrorBoundary";
 
 export { ServerSelectBox } from "./ServerSelectBox";
+export { Counter } from "./Counter";
 export { ProductList } from "./ProductList";
 export { CartProductList } from "./CartProductList";
 export { OrderProductList } from "./OrderProductList";
